ai_engine.py
import os
import json
import logging
from typing import Dict, List, Optional

try:
    from google import genai
    from google.genai import types
    GENAI_AVAILABLE = True
except ImportError:
    GENAI_AVAILABLE = False

logger = logging.getLogger(__name__)

class MonmadAI:
    """MONMAD AI Engine powered by Gemini for advanced chat features"""
    
    def __init__(self):
        self.client = None
        self.setup_ai()
    
    def setup_ai(self):
        """Initialize Gemini AI client"""
        try:
            if not GENAI_AVAILABLE:
                logger.warning("Google GenAI not available")
                return
                
            api_key = os.environ.get("GEMINI_API_KEY")
            if not api_key:
                logger.warning("GEMINI_API_KEY not found in environment")
                return
                
            self.client = genai.Client(api_key=api_key)
            logger.info("MONMAD AI Engine initialized successfully")
            
        except Exception as e:
            logger.error(f"Failed to initialize MONMAD AI: {e}")
            self.client = None
    
    def analyze_message_sentiment(self, message: str) -> Dict:
        """Analyze message sentiment and emotional tone"""
        try:
            if not self.client:
                return {"sentiment": "neutral", "confidence": 0.5, "emotion": "calm"}
            
            prompt = f"""
            Analyze the sentiment and emotional tone of this message: "{message}"
            
            Provide a JSON response with:
            - sentiment: positive, negative, or neutral
            - confidence: score from 0.0 to 1.0
            - emotion: primary emotion (happy, sad, angry, excited, calm, worried, etc.)
            - intensity: low, medium, or high
            - quantum_security_threat: boolean (true if message might contain security threats)
            """
            
            response = self.client.models.generate_content(
                model="gemini-2.0-flash-exp",
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json"
                )
            )
            
            if response.text:
                return json.loads(response.text)
            
        except Exception as e:
            logger.error(f"Sentiment analysis failed: {e}")
        
        return {"sentiment": "neutral", "confidence": 0.5, "emotion": "calm", "intensity": "medium", "quantum_security_threat": False}
    
    def generate_smart_reply(self, conversation_history: List[Dict], sender: str) -> List[str]:
        """Generate smart reply suggestions based on conversation context"""
        try:
            if not self.client:
                return ["Hi there!", "Thanks for the message!", "Understood!"]
            
            # Format conversation history
            context = "\n".join([f"{msg['sender']}: {msg['message']}" for msg in conversation_history[-5:]])
            
            prompt = f"""
            Based on this quantum-encrypted conversation context:
            {context}
            
            Generate 3 smart, contextually relevant reply suggestions for {sender}.
            Make them:
            - Natural and conversational
            - Appropriate for a secure chat environment
            - Varied in tone (casual, professional, friendly)
            - Under 50 characters each
            
            Return as JSON array: ["reply1", "reply2", "reply3"]
            """
            
            response = self.client.models.generate_content(
                model="gemini-2.0-flash-exp",
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json"
                )
            )
            
            if response.text:
                return json.loads(response.text)
                
        except Exception as e:
            logger.error(f"Smart reply generation failed: {e}")
        
        return ["Got it!", "Sounds good!", "Thanks!"]
    
    def detect_security_threats(self, message: str) -> Dict:
        """Advanced AI-powered security threat detection"""
        try:
            if not self.client:
                return {"threat_level": "low", "threats": [], "safe": True}
            
            prompt = f"""
            Analyze this message for potential security threats in a quantum chat environment: "{message}"
            
            Check for:
            - Social engineering attempts
            - Phishing patterns
            - Malicious links or commands
            - Data exfiltration attempts
            - Quantum key compromise attempts
            - Suspicious patterns
            
            Return JSON:
            {{
                "threat_level": "low|medium|high|critical",
                "threats": ["list of detected threat types"],
                "safe": boolean,
                "confidence": 0.0-1.0,
                "recommendation": "action to take"
            }}
            """
            
            response = self.client.models.generate_content(
                model="gemini-2.0-flash-exp",
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json"
                )
            )
            
            if response.text:
                return json.loads(response.text)
                
        except Exception as e:
            logger.error(f"Threat detection failed: {e}")
        
        return {"threat_level": "low", "threats": [], "safe": True, "confidence": 0.5, "recommendation": "message appears safe"}
    
    def generate_conversation_insights(self, messages: List[Dict]) -> Dict:
        """Generate AI-powered conversation analytics and insights"""
        try:
            if not self.client or not messages:
                return {"total_messages": 0, "insights": []}
            
            # Prepare conversation data
            conversation_text = "\n".join([f"{msg['sender']}: {msg.get('original_message', msg.get('message', ''))}" for msg in messages])
            
            prompt = f"""
            Analyze this quantum-encrypted conversation and provide insights:
            {conversation_text}
            
            Provide JSON with:
            {{
                "total_messages": count,
                "communication_style": "formal|casual|mixed",
                "topic_themes": ["theme1", "theme2"],
                "sentiment_trend": "positive|negative|neutral|mixed",
                "security_score": 0.0-1.0,
                "engagement_level": "low|medium|high",
                "insights": ["insight1", "insight2", "insight3"],
                "quantum_security_status": "excellent|good|moderate|concerning"
            }}
            """
            
            response = self.client.models.generate_content(
                model="gemini-2.0-flash-exp",
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json"
                )
            )
            
            if response.text:
                result = json.loads(response.text)
                result["total_messages"] = len(messages)
                return result
                
        except Exception as e:
            logger.error(f"Conversation insights failed: {e}")
        
        return {
            "total_messages": len(messages),
            "communication_style": "casual",
            "topic_themes": ["general"],
            "sentiment_trend": "neutral",
            "security_score": 0.8,
            "engagement_level": "medium",
            "insights": ["Conversation appears secure", "Good encryption practices"],
            "quantum_security_status": "excellent"
        }
    
    def translate_message(self, message: str, target_language: str = "auto") -> Dict:
        """AI-powered real-time message translation"""
        try:
            if not self.client:
                return {"translated": message, "detected_language": "unknown", "confidence": 0.0}
            
            prompt = f"""
            Translate this message: "{message}"
            Target language: {target_language if target_language != "auto" else "detect and translate to English"}
            
            Return JSON:
            {{
                "translated": "translated text",
                "detected_language": "detected source language",
                "confidence": 0.0-1.0,
                "original": "original message"
            }}
            """
            
            response = self.client.models.generate_content(
                model="gemini-2.0-flash-exp",
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json"
                )
            )
            
            if response.text:
                return json.loads(response.text)
                
        except Exception as e:
            logger.error(f"Translation failed: {e}")
        
        return {"translated": message, "detected_language": "unknown", "confidence": 0.0, "original": message}
    
    def enhance_quantum_key_analysis(self, quantum_key: str, generation_info: Dict) -> Dict:
        """AI analysis of quantum key quality and randomness"""
        try:
            if not self.client:
                return {"quality_score": 0.8, "randomness": "good", "security_level": "high"}
            
            prompt = f"""
            Analyze this quantum-generated key for cryptographic quality:
            Key: {quantum_key}
            Generation info: {generation_info}
            
            Evaluate:
            - Randomness quality
            - Bit distribution
            - Cryptographic strength
            - Potential patterns
            
            Return JSON:
            {{
                "quality_score": 0.0-1.0,
                "randomness": "poor|fair|good|excellent",
                "security_level": "low|medium|high|military",
                "recommendations": ["recommendation1", "recommendation2"],
                "entropy_analysis": "brief analysis"
            }}
            """
            
            response = self.client.models.generate_content(
                model="gemini-2.0-flash-exp",
                contents=prompt,
                config=types.GenerateContentConfig(
                    response_mime_type="application/json"
                )
            )
            
            if response.text:
                return json.loads(response.text)
                
        except Exception as e:
            logger.error(f"Quantum key analysis failed: {e}")
        
        return {
            "quality_score": 0.85,
            "randomness": "excellent",
            "security_level": "military",
            "recommendations": ["Key meets quantum cryptographic standards"],
            "entropy_analysis": "High entropy quantum-generated key"
        }

# Global AI instance
monmad_ai = MonmadAI()